[PY] Allow direct import of ripser_parallel and move test folders (#36)
* Allow direct import of ripser_parallel, move python test folders inside python dir

Fixes #15
