+++ gph/python/__init__.py
@@ -1,1 +1,5 @@
 from .ripser_interface import ripser_parallel
+
+__all__ = [
+    "ripser_parallel",
+]
