+++ gph/python/test/test_collapser.py
